refactor(mailbox): remove dead actions and derive MCP tools
