+++ openspec/changes/derive-mcp-tools/tasks.md
@@ -1,0 +1,10 @@
+# Tasks
+
+## 1. Derive tools
+- [x] 1.1 Replace the hand-maintained wrapped-action list with a derivation from `EMAIL_ACTIONS`.
+- [x] 1.2 Preserve state-aware overrides for read/list/search and mailbox diagnostics.
+- [x] 1.3 Assert intentional exclusions and duplicate-free tool names.
+
+## 2. Verification
+- [x] 2.1 Add registry-to-tool parity coverage.
+- [x] 2.2 Run tests, lint, build, strict OpenSpec validation, and spec coverage.
